Change signature of remoteSendStreamData() to fix compile warning
The actual type of size_t is architecture dependent. Because the len
parameter is used as unsigned int in remoteSendStreamData(), change its
type to unsigned int.

* daemon/dispatch.[ch]: change size_t to unsigned int for
  remoteSendStreamData()

+++ daemon/dispatch.h
@@ -75,6 +75,6 @@ int
 remoteSendStreamData(struct qemud_client *client,
                      struct qemud_client_stream *stream,
                      const char *data,
-                     size_t len);
+                     unsigned int len);
 
 #endif /* __LIBVIRTD_DISPATCH_H__ */
